+++ resources/js/api/Follow.js
@@ -1,4 +1,4 @@
-import API from "../api/api";
+import API from "./api";
 
 export class Follow {
